Update arrCopy and arrInit functions to use unsigned long for size parameter

+++ Solver/Kernels/mathfunc.h
@@ -1,6 +1,5 @@
 #ifndef MATHFUNC_H
 #define MATHFUNC_H
-#include <vector>
 
 void dot(const double *x, const double *y, int size, double &res);
 
@@ -21,7 +20,7 @@ void axpy(double a, const double *x, const double *y, int size, double *res);
  * @param size array size
  */
 template <typename T>
-void arrCopy(T *y, const T *x, const int size) {
+void arrCopy(T *y, const T *x, const unsigned long size) {
     #pragma omp parallel for default(none) shared(x, y, size)
     for (int i = 0; i < size; i++) {
         y[i] = x[i];
@@ -38,7 +37,7 @@ void arrCopy(T *y, const T *x, const int size) {
  * @param size array size
  */
 template <typename T>
-void arrInit(T *x, const T a, const int size) {
+void arrInit(T *x, const T a, const unsigned size) {
     #pragma omp parallel for default(none) shared(x, a, size)
     for (int i = 0; i < size; i++) {
         x[i] = a;
